feat: Moves sound notification manager in a separate package.
